Objetos del jugador, cambios en el menú

+++ Assets/Prefabs/Scripts/BarraArmadura.cs
@@ -6,11 +6,13 @@ using UnityEngine.UI;
 public class BarraArmadura : MonoBehaviour
 {
     public Slider slider;
-    public void SetMaxArmor(float armor){
+    //Mediante el valor máximo, obtenemos el actual
+    public void SetMaxArmor(int armor){
         slider.maxValue = armor;
         slider.value = armor/2;
     }
-    public void SetArmor(float armor){
+    //Actualizamos la vida
+    public void SetArmor(int armor){
         slider.value = armor;
     }
 }
